refactor because main is getting way too large
